Extract symoblic machine generator.
