+++ autoprecompiles/src/lib.rs
@@ -1,9 +1,7 @@
 use constraint_optimizer::IsBusStateful;
 use itertools::Itertools;
-use powdr::{Column, UniqueColumns};
-use powdr_ast::analyzed::{
-    AlgebraicBinaryOperation, AlgebraicExpression, AlgebraicReference, AlgebraicUnaryOperator,
-};
+use powdr::UniqueColumns;
+use powdr_ast::analyzed::{AlgebraicExpression, AlgebraicReference, AlgebraicUnaryOperator};
 use powdr_ast::analyzed::{PolyID, PolynomialType};
 use powdr_ast::parsed::visitor::Children;
 use powdr_constraint_solver::constraint_system::BusInteractionHandler;
@@ -11,6 +9,7 @@ use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 use std::fmt::Display;
 use std::iter::once;
+use symbolic_machine_generator::statements_to_symbolic_machine;
 
 use powdr_number::{FieldElement, LargeInt};
 use powdr_pilopt::simplify_expression;
@@ -19,6 +18,7 @@ pub mod constraint_optimizer;
 pub mod optimizer;
 pub mod powdr;
 pub mod register_optimizer;
+pub mod symbolic_machine_generator;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct SymbolicInstructionStatement<T> {
@@ -279,9 +279,9 @@ pub enum VMBusInteraction<T> {
     Memory(MemoryBusInteraction<T>),
 }
 
-const EXECUTION_BUS_ID: u64 = 0;
-const MEMORY_BUS_ID: u64 = 1;
-const PC_LOOKUP_BUS_ID: u64 = 2;
+pub const EXECUTION_BUS_ID: u64 = 0;
+pub const MEMORY_BUS_ID: u64 = 1;
+pub const PC_LOOKUP_BUS_ID: u64 = 2;
 
 pub fn build<T: FieldElement>(
     program: Vec<SymbolicInstructionStatement<T>>,
@@ -292,7 +292,7 @@ pub fn build<T: FieldElement>(
     opcode: u32,
 ) -> (SymbolicMachine<T>, Vec<Vec<u64>>) {
     let (machine, subs) =
-        generate_symbolic_machine(&program, &instruction_kind, &instruction_machines);
+        statements_to_symbolic_machine(&program, &instruction_kind, &instruction_machines);
 
     let machine = optimizer::optimize(machine, bus_interaction_handler, opcode, degree_bound);
 
@@ -396,186 +396,4 @@ fn add_guards<T: FieldElement>(mut machine: SymbolicMachine<T>) -> SymbolicMachi
     machine.constraints.extend(is_valid_mults);
 
     machine
-}
-
-fn exec_receive<T: FieldElement>(machine: &SymbolicMachine<T>) -> SymbolicBusInteraction<T> {
-    let [r, _s] = machine
-        .bus_interactions
-        .iter()
-        .filter_map(|bus_int| match bus_int.id {
-            EXECUTION_BUS_ID => Some(bus_int.clone()),
-            _ => None,
-        })
-        .collect::<Vec<_>>()
-        .try_into()
-        .unwrap();
-    // TODO assert that r.mult matches -expr
-    r
-}
-
-pub fn generate_symbolic_machine<T: FieldElement>(
-    statements: &[SymbolicInstructionStatement<T>],
-    instruction_kinds: &BTreeMap<String, InstructionKind>,
-    instruction_machines: &BTreeMap<String, SymbolicMachine<T>>,
-) -> (SymbolicMachine<T>, Vec<Vec<u64>>) {
-    let mut constraints: Vec<SymbolicConstraint<T>> = Vec::new();
-    let mut bus_interactions: Vec<SymbolicBusInteraction<T>> = Vec::new();
-    let mut col_subs: Vec<Vec<u64>> = Vec::new();
-    let mut global_idx: u64 = 3;
-
-    for (i, instr) in statements.iter().enumerate() {
-        match instruction_kinds.get(&instr.name).unwrap() {
-            InstructionKind::Normal
-            | InstructionKind::UnconditionalBranch
-            | InstructionKind::ConditionalBranch => {
-                let machine = instruction_machines.get(&instr.name).unwrap().clone();
-
-                let (next_global_idx, subs, machine) = powdr::reassign_ids(machine, global_idx, i);
-                global_idx = next_global_idx;
-
-                let pc_lookup: PcLookupBusInteraction<T> = machine
-                    .bus_interactions
-                    .iter()
-                    .filter_map(|bus_int| bus_int.clone().try_into().ok())
-                    .exactly_one()
-                    .expect("Expected single pc lookup");
-
-                let mut sub_map: BTreeMap<Column, AlgebraicExpression<T>> = Default::default();
-                let mut local_constraints: Vec<SymbolicConstraint<T>> = Vec::new();
-
-                let is_valid: AlgebraicExpression<T> = exec_receive(&machine).mult.clone();
-                let one = AlgebraicExpression::Number(1u64.into());
-                local_constraints.push((is_valid.clone() + one).into());
-
-                let mut sub_map_loadstore: BTreeMap<Column, AlgebraicExpression<T>> =
-                    Default::default();
-                if is_loadstore(instr.opcode) {
-                    sub_map_loadstore.extend(loadstore_chip_info(&machine, instr.opcode));
-                }
-
-                // Constrain the opcode expression to equal the actual opcode.
-                let opcode_constant = AlgebraicExpression::Number((instr.opcode as u64).into());
-                local_constraints.push((pc_lookup.op.clone() - opcode_constant).into());
-
-                assert_eq!(instr.args.len(), pc_lookup.args.len());
-                instr
-                    .args
-                    .iter()
-                    .zip_eq(&pc_lookup.args)
-                    .for_each(|(instr_arg, pc_arg)| {
-                        let arg = AlgebraicExpression::Number(*instr_arg);
-                        match pc_arg {
-                            AlgebraicExpression::Reference(ref arg_ref) => {
-                                sub_map.insert(Column::from(arg_ref), arg);
-                            }
-                            AlgebraicExpression::BinaryOperation(_expr) => {
-                                local_constraints.push((arg - pc_arg.clone()).into());
-                            }
-                            AlgebraicExpression::UnaryOperation(_expr) => {
-                                local_constraints.push((arg - pc_arg.clone()).into());
-                            }
-                            _ => {}
-                        }
-                    });
-
-                let local_identities = machine
-                    .constraints
-                    .iter()
-                    .chain(&local_constraints)
-                    .map(|expr| {
-                        let mut expr = expr.expr.clone();
-                        powdr::substitute_algebraic(&mut expr, &sub_map);
-                        powdr::substitute_algebraic(&mut expr, &sub_map_loadstore);
-                        expr = simplify_expression(expr);
-                        SymbolicConstraint { expr }
-                    })
-                    .collect::<Vec<_>>();
-
-                constraints.extend(local_identities);
-
-                for bus_int in &machine.bus_interactions {
-                    let mut link = bus_int.clone();
-                    link.args
-                        .iter_mut()
-                        .chain(std::iter::once(&mut link.mult))
-                        .for_each(|e| {
-                            powdr::substitute_algebraic(e, &sub_map);
-                            powdr::substitute_algebraic(e, &sub_map_loadstore);
-                            *e = simplify_expression(e.clone());
-                        });
-                    bus_interactions.push(link);
-                }
-
-                col_subs.push(subs);
-
-                // after the first round of simplifying,
-                // we need to look for register memory bus interactions
-                // and replace the addr by the first argument of the instruction
-                for bus_int in &mut bus_interactions {
-                    if bus_int.id != MEMORY_BUS_ID {
-                        continue;
-                    }
-
-                    let addr_space = match bus_int.args[0] {
-                        AlgebraicExpression::Number(n) => n.to_integer().try_into_u32().unwrap(),
-                        _ => panic!(
-                            "Address space must be a constant but got {}",
-                            bus_int.args[0]
-                        ),
-                    };
-
-                    if addr_space != 1 {
-                        continue;
-                    }
-
-                    match bus_int.args[1] {
-                        AlgebraicExpression::Number(_) => {}
-                        _ => {
-                            if let Some(arg) = bus_int.args.get_mut(1) {
-                                *arg = instr.args[0].into();
-                            } else {
-                                panic!("Expected address argument");
-                            }
-                        }
-                    };
-                }
-            }
-            _ => {}
-        }
-    }
-
-    (
-        SymbolicMachine {
-            constraints,
-            bus_interactions,
-        },
-        col_subs,
-    )
-}
-
-fn is_loadstore(opcode: usize) -> bool {
-    (0x210..=0x215).contains(&opcode)
-}
-
-fn loadstore_chip_info<T: FieldElement>(
-    machine: &SymbolicMachine<T>,
-    opcode: usize,
-) -> BTreeMap<Column, AlgebraicExpression<T>> {
-    let is_load = if opcode == 0x210 || opcode == 0x211 || opcode == 0x212 {
-        T::from(1u32)
-    } else {
-        T::from(0u32)
-    };
-    let is_load = AlgebraicExpression::Number(is_load);
-    let is_load_expr = match &machine.constraints[7].expr {
-        AlgebraicExpression::BinaryOperation(AlgebraicBinaryOperation { left, .. }) => left.clone(),
-        _ => panic!("Expected subtraction."),
-    };
-    let is_load_col = if let AlgebraicExpression::Reference(r) = &*is_load_expr {
-        r.into()
-    } else {
-        panic!("expected a single reference")
-    };
-
-    [(is_load_col, is_load)].into()
 }
